refactor: split up into several packages
Add whatever glue code is needed for compilation.

fix: Add some quick hacks for compilation

The 'pure' library has since moved on to version 5.0.5 while the code
here is written for version 4.2.0: a breaking change.
The 'uuid' library is strange, the source code for that package is not
on go modules, but I think the go.mod file implies that it is (doesn't
have the words "incompatible" next to it in go.mod). Anyway, the weird
thing is that the NewV4 function returns (uuid.UUID, error), but when I
try to compile code, it says that only 1 value is returned.
Consider those changes to be workarounds.