feat: Start implementing extension jobs after user item syncs
This is adapted from the ruby version of standardnotes sync server.

refactor: Move high-level code out of models

Attempting to extract application logic out of models package and into
ineractors package. The goal is to reduce models to database access,
model entities. In the future, that may be split some more.