refactor: Clean up models; rm Loadable interface
For more uniformity, use pointer receivers in most methods. However,
there are a few times that value receivers are a better choice, such as
the MakeSaferCopy, Copy methods, which are simpler as a value receiver.
Changed signature of LoadByUUID and Exists methods, b/c we should pay
attention to the error as an error. Remove unused interface, redundant
logging and methods.

refactor: Move, consolidate models.User methods

refactor: remove Loadable interface

Did not see it used anywhere and I don't find it useful.