refactor: Simplify item conflict handling
I don't think a set is necessary in the previous implementation. Sure,
it's a threadsafe set (I checked source code), but this is synchronous
code. It does two linear Finds for each conflict anyways, which will
likely throw away all the constant time benefits of having a set. A
regular map is better for this purpose since it can be used to lookup
conflicted items in the loop.

refactor: Begin consolidating sync interactor functions

There is 1 less function than before, naming is easier. Need to update
the conflict resolution stuff and sync response fields.

refactor: Clean up conflict code some more

It should be based closely off of the rails server implementation of
*standardnotes*. Decided to throw one of the conflict functions,
checkHandleConflicts, since that was based on an older version of
standardfile (which is deprecated itself). I've checked the client side
code and can confirm that it's looking for a key named "conflicts".
Though the older, "unsaved" key still works, I'm not going to worry
about that.

refactor: Move sync types, improve sync item & conflict code