fix: Pay attention to errors and make loading less awkward
fix: Update db.SelectStruct function

This function is used for selecting one row out of DB. It was not
signaling errors, such as sql.NoRowsFound. Unfortunately, cannot use the
sql.QueryRows function from the stdlib b/c of the way sqlstruct library
works. Added a workaround in order to manually return it but ignore
subsequent rows in the query. This is not ideal because there are some
functions that lookup users by email, which doesn't have a unique
constraint! I think that issue is better addressed by either adding the
unique constraint or changing the index to have the constraint; at a
later time.

refactor: change the signature to only return an error. There was no
point in returning the same value you're passing in to the function.
Instead, expect the input to be mutated by the db query.

refactor: Make Item, User loading from db less awkward

It was awkward because you needed a pointer to a User or Item before you
could fetch anything else from the DB. Now just use UUID. Also make use
of the no rows found error.