refactor: Clean up models, remove pw_auth db col, rename Params type
Previous implementation of models.Item.Update method did not update the
ContentType field. Following the Rails version implementation, which
allows that, however the progenitor golang version did not.
The pw_auth column is not in StandardNotes, it was removed over 2 years
ago. I think it was added in the golang version before that. This is
pretty much a complete rewrite, so removing it from migrations as well.

refactor: Combine User update methods, move pw updates to interactor

I think having a separate method for updating the password and the other
parts of the User led to more confusion. The UpdateParams method took a
misnamed (I misnamed it earlier in a refactor) argument and didn't use
it at all. It was misnamed b/c I thought GET auth/params should output a
struct, rather than a map, so I combined that with Params. In the
original implementation, Params serves as named parameters to update a
User (just not the password, for that you use NewPassword). Confused
yet? That's why the User methods and support types are being reworked.

Since the UpdateParams and UpdatePassword method are now combined into
one Update method, moved the User password update stuff to an
interactor, as that is a high-level concern IMO.

refactor: Rename Params type to PwGenParams

The name Params is too vague, has caused much confusion.

docs: Update godoc for models