fix: Add various itemsync-related fixes
fix: Apply itemsync pagination fixes

The client was spamming server because it was not getting the correct
signals, in the form of tokens. The protocol says to add a cursor_token
when there are more results, and the token decodes to a timestamp. The
problem: the server was either always adding or omitting the token.
Whenever the token was included, it was of the earliest result, whereas
it should have been the latest.

Improve the internal heuristic to ascertain if there are results. The
only change needed there was to compare using `>=` instead of `>`.

Fix a panic due to nil reference.

refactors:

Assigning the output token to the itemsync.Response was spread out
between two functions, so move all of that stuff into one. Make method
and variable names clearer. Some other changes: sort ASC, add result
limits.