feat: Add db interfaces, query types; rm db migrations code
The goal is to make it easier to query for collections of items and know
if there are more results or not. The current db query functions use a
lot of complicated runtime checks (in the reflect package) that I don't
understand, so I want to code it out.

Remove db function, Select.
Fix jobs code, which used the db.Select function to use the existing
methods for models.User. NOTE: The item fetching function adds one more
condition to the query but I don't think it matters that much.

refactor: remove db migrations code

Just thought about this file for the first time in weeks. It is a broken
dependent of the previous removal of the db.Select function. I don't
think this code is even necessary, because the db column is already in
the schema and there are no users. This of course refers to any running
instances of this StandardNotes implementation, of which there are none.
Can also remove the MigratingUser implementation for models.User.

build: run go mod tidy

removed the migration dependency