fix: Clean up itemsync pagination issues; fix conflict problem
The pagination token had the local timezone, this may have caused issues
with the client, which operates in UTC. Everything timestamp in this
application should be UTC wherever possible.

Also make use of the updates to User item loading functions, it provides
a better signal of more results.

The problem of constant item conflicts was that the unit used to
compare timestamps was way off. There was no need multiply the
difference by time.Seconds, the other operand was already in the same
unit as the difference.