build: Replace sqlite3 with mysql
build: Update Makefile with db, db migration commands
build: Update db schema, add migrations
build: Add docker
build: Add db migration tool

fix: db query related issues in models

We're not going to store datetime values at a precision finer than 6
radix points (storing microseconds, but not nanoseconds). The width of a
datetime on its own is 5 bytes, and add 3 more for fractional any
seconds. Going over 8 would likely cause extra cache-alignment costs
for precision we don't need for this application. Also, it may not even
be supported in MySQL, I don't know. Either way, it's too fine.

In golang, the in-memory time.Time of a struct *before* it's saved to
the DB would be in nanoseconds. Loading that same struct from the
database gives you microseconds. This mismatch (object relation
impedance mismatch?) caused some problems in the models tests so a
time comparison function was introduced. The mismatch also caused issues
in the itemsync, but was fixed by rounding.

test: Use measured non-determinism to make tests pass

Uniqueness constraints that get in the way of passing tests now. When
using sqlite's in-memory database features, the context was thrown away
for each db connection, so it wasn't a problem then.

fix: Increase width of db column & update docker files

The enc_item_key is somewhere around 560 bytes. The current limit is a
square number and leaves room. The old limit (255) was present when the
application was using sqlite; I don't know if that mattered then.
Also update items.content to longblob, which can store up to 4 GB (blob
stores up to 64 KB). A mediumblob store 16 MB, which may not be enough.
As for going with a blob type rather than text, I don't think it matters
a whole lot since we won't be searching or sorting on this column. The
text column types use character strings, and can use a specific
collation. Blob is binary, this column is encrypted and opaque.

build: Update docker files

Add some missing things to get it working. When it comes to
cross-container network communication using docker-compose, I *think*
you can reference another container's port by mentioning the service's
name in the docker-compose file? This may prove to be a fluke.